+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Schema/PbrSpecularGlossiness.cs
@@ -14,23 +14,12 @@ namespace GLTFast.Schema
     /// <seealso href="https://github.com/KhronosGroup/glTF/tree/main/extensions/2.0/Archived/KHR_materials_pbrSpecularGlossiness"/>
     public class PbrSpecularGlossiness
     {
-
         /// <summary>
         /// Diffuse color red, green, blue and alpha components in linear space.
         /// </summary>
         [JsonPropertyName("diffuseFactor")]
-        public float[] DiffuseFactor { get; set; } = { 1, 1, 1, 1 };
-
-        /// <summary>
-        /// Diffuse color in linear space.
-        /// </summary>
-        public Color DiffuseColor =>
-            new Color(
-                DiffuseFactor[0],
-                DiffuseFactor[1],
-                DiffuseFactor[2],
-                DiffuseFactor[3]
-            );
+        [JsonConverter(typeof(ColorAlphaConverter))]
+        public ColorAlpha DiffuseFactor { get; set; } = ColorAlpha.White;
 
         /// <summary>
         /// Diffuse color texture info.
@@ -42,17 +31,8 @@ namespace GLTFast.Schema
         /// Specular color red, green and blue components in linear space.
         /// </summary>
         [JsonPropertyName("specularFactor")]
-        public float[] SpecularFactor { get; set; } = { 1, 1, 1 };
-
-        /// <summary>
-        /// Specular color in linear space.
-        /// </summary>
-        public Color SpecularColor =>
-            new Color(
-                SpecularFactor[0],
-                SpecularFactor[1],
-                SpecularFactor[2]
-            );
+        [JsonConverter(typeof(ColorConverter))]
+        public Color SpecularFactor { get; set; } = Color.White;
 
         /// <summary>
         /// The glossiness or smoothness of the material.